+++ src/plugins_types/enumeration.c
@@ -37,7 +37,7 @@
  */
 
 LIBYANG_API_DEF LY_ERR
-lyplg_type_store_enum(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value, size_t value_len,
+lyplg_type_store_enum(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value, uint32_t value_len,
         uint32_t options, LY_VALUE_FORMAT format, void *UNUSED(prefix_data), uint32_t hints,
         const struct lysc_node *UNUSED(ctx_node), struct lyd_value *storage, struct lys_glob_unres *UNUSED(unres),
         struct ly_err_item **err)
@@ -56,8 +56,8 @@ lyplg_type_store_enum(const struct ly_ctx *ctx, const struct lysc_type *type, co
     if (format == LY_VALUE_LYB) {
         /* validation */
         if (value_len != 4) {
-            ret = ly_err_new(err, LY_EVALID, LYVE_DATA, NULL, NULL, "Invalid LYB enumeration value size %zu (expected 4).",
-                    value_len);
+            ret = ly_err_new(err, LY_EVALID, LYVE_DATA, NULL, NULL, "Invalid LYB enumeration value size %" PRIu32
+                    " (expected 4).", value_len);
             goto cleanup;
         }
 
@@ -149,7 +149,7 @@ lyplg_type_sort_enum(const struct ly_ctx *UNUSED(ctx), const struct lyd_value *v
 
 LIBYANG_API_DEF const void *
 lyplg_type_print_enum(const struct ly_ctx *UNUSED(ctx), const struct lyd_value *value, LY_VALUE_FORMAT format,
-        void *UNUSED(prefix_data), ly_bool *dynamic, size_t *value_len)
+        void *UNUSED(prefix_data), ly_bool *dynamic, uint32_t *value_len)
 {
     int64_t prev_num = 0, num = 0;
     void *buf;
